+++ default.nix
@@ -2,8 +2,8 @@
 (import ./reflex-platform {}).project ({ pkgs, ... }: {
   packages = {
     common = ./common;
-    # backend = ./backend;
-    /* frontend = ./frontend; */
+    backend = ./backend;
+    frontend = ./frontend;
     basictodo = ./BasicTodo;
     draganddrop = ./drag-and-drop;
     fileinput = ./fileinput;
@@ -14,8 +14,10 @@
     websocketecho = ./websocket-echo;
     xhrblob = ./xhr-blob;
     wschat = ./ws-chat;
-    /* commonwschat = ./common-ws-chat; */
+    wschat2 = ./ws-chat2;
+    commonwschat2 = ./common-ws-chat2;
     backendwschat = ./backend-ws-chat;
+    backendwschat2 = ./backend-ws-chat2;
   };
 
   /* android.frontend = { */
@@ -34,13 +36,13 @@
     ghc = ["keyboard" "basictodo" "draganddrop" "fileinput" "nasapod"
         "othello" "pegsolitaire" "websocketecho" "xhrblob"
         "wschat" "backendwschat"
+        "wschat2" "backendwschat2" "commonwschat2"
     ];
     ghcjs = ["keyboard" "basictodo" "draganddrop" "fileinput" "nasapod"
         "othello" "pegsolitaire" "websocketecho" "xhrblob"
         "wschat"
+        "wschat2" "commonwschat2"
     ];
-    /* ghc = ["frontend"]; */
-    /* ghcjs = ["frontend"]; */
     /* ghc = ["common" "backend" "frontend"]; */
     /* ghcjs = ["common" "frontend"]; */
   };