+++ ChangeLog.md
@@ -6,7 +6,7 @@
 * Use "project" from reflex-platform for all examples.
 * Small fixes to nasa-pod -example.
 * Small fixes to drag-and-drop -example.
-* Added a simple websocket chat example.
+* Added two simple websocket chat examples.
 * Changed .gitignore a bit
 * Other minor changes.